feat: add modules section task solution
